requirements.txt
altgraph==0.15
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.2
bs4==0.0.1
click==6.7
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
Flask==0.11.1
Flask-SQLAlchemy==2.3.2
future==0.16.0
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
h5py==2.7.1
html5lib==1.0.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
macholib==1.9
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pefile==2017.11.5
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyInstaller==3.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
scipy==1.0.0
selenium==3.11.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.6
terminado==0.8.1
testpath==0.3.1
torch==0.3.1
torchvision==0.2.0
tornado==4.5.3
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.3
